+++ src/components/radio-group/radio-group.ts
@@ -9,6 +9,7 @@ import { WebAwesomeFormAssociatedElement } from '../../internal/webawesome-form-
 import formControlStyles from '../../styles/shadow/form-control.css';
 import buttonGroupStyles from '../../styles/utilities/button-group.css';
 import sizeStyles from '../../styles/utilities/size.css';
+import '../radio-button/radio-button.js';
 import type WaRadioButton from '../radio-button/radio-button.js';
 import '../radio/radio.js';
 import type WaRadio from '../radio/radio.js';
@@ -20,7 +21,8 @@ import styles from './radio-group.css';
  * @status stable
  * @since 2.0
  *
- * @dependency wa-button-group
+ * @dependency wa-radio
+ * @dependency wa-radio-button
  *
  * @slot - The default slot where `<wa-radio>` or `<wa-radio-button>` elements are placed.
  * @slot label - The radio group's label. Required for proper accessibility. Alternatively, you can use the `label`
@@ -34,6 +36,7 @@ import styles from './radio-group.css';
  * @csspart form-control - The form control that wraps the label, input, and hint.
  * @csspart form-control-label - The label's wrapper.
  * @csspart form-control-input - The input's wrapper.
+ * @csspart radios - The wrapper than surrounds radio items, styled as a flex container by default.
  * @csspart hint - The hint's wrapper.
  * @csspart button-group - The button group that wraps radio buttons.
  * @csspart button-group__base - The button group's `base` part.
@@ -63,6 +66,7 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
   @query('slot:not([name])') defaultSlot: HTMLSlotElement;
 
   @state() private hasButtonGroup = false;
+  @state() private hasRadioButtons = false;
 
   /**
    * The radio group's label. Required for proper accessibility. If you need to display HTML, use the `label` slot
@@ -75,6 +79,9 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
 
   /** The name of the radio group, submitted as a name/value pair with form data. */
   @property({ reflect: true }) name: string | null = null;
+
+  /** The orientation in which to show radio items. */
+  @property({ reflect: true }) orientation: 'horizontal' | 'vertical' = 'vertical';
 
   private _value: string | null = null;
 
@@ -172,6 +179,9 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
 
   private async syncRadioElements() {
     const radios = this.getAllRadios();
+
+    // Detect the presence of radio buttons
+    this.hasRadioButtons = radios.some(radio => radio.localName === 'wa-radio-button');
 
     await Promise.all(
       // Sync the checked state and size
@@ -324,11 +334,13 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
           'form-control--large': this.size === 'large',
           'form-control--radio-group': true,
           'form-control--has-label': hasLabel,
+          'form-control--has-radio-buttons': this.hasRadioButtons,
         })}
         role="radiogroup"
         aria-labelledby="label"
         aria-describedby="hint"
         aria-errormessage="error-message"
+        aria-orientation=${this.orientation}
       >
         <label
           part="form-control-label"
@@ -342,7 +354,10 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
 
         <slot
           part="form-control-input"
-          class=${classMap({ 'wa-button-group': this.hasButtonGroup })}
+          class=${classMap({
+            'wa-button-group': this.hasButtonGroup,
+            'wa-button-group-vertical': this.orientation === 'vertical',
+          })}
           @slotchange=${this.syncRadioElements}
         ></slot>